Change Menu UI & Splash UI & Welcome Activities & OrderActivity.kt

+++ app/src/main/java/com/banquemisr/coffeeapp_banquemisr/domain/model/CoffeeOrder.kt
@@ -15,7 +15,7 @@ data class CoffeeOrder(
     val count: Int?,
     val size: String?,
     val sugar: String?,
-    val totalPrice: Float?,
+    val totalPrice: Int?,
     val icon: Int,
 ) : Serializable
